Backend/requirements.txt
Flask>=1.0.2
torch>=1.11.0
torchvision>=0.12.0
numpy
Pillow>=9.0.0
gunicorn>=20.1.0
wget>=3.2
aiofiles>=0.8.0
aiohttp>=3.8.1
uvicorn>=0.17.6
nest_asyncio>=1.5.5
requests>=2.25.1
PyYAML>=5.4.1
gdown
Flask-Cors>=3.0.10
